+++ src/main/java/com/webtoon/content/dto/request/ContentSave.java
@@ -18,14 +18,14 @@ public class ContentSave {
     private String subTitle;
 
     @Min(value = 0, message = "몇 화인지를 입력해주세요")
-    private Integer episode;
+    private Long episode;
 
     @NotNull(message = "등록 날짜를 입력해주세요")
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private LocalDate registrationDate;
 
     @Builder
-    public ContentSave(String subTitle, Integer episode, LocalDate registrationDate) {
+    public ContentSave(String subTitle, Long episode, LocalDate registrationDate) {
         this.subTitle = subTitle;
         this.episode = episode;
         this.registrationDate = registrationDate;
