+++ src/test/java/com/webtoon/cartoon/service/CartoonServiceTest.java
@@ -174,7 +174,7 @@ class CartoonServiceTest extends ServiceTest {
                 .genre("NONE")
                 .build();
 
-        List<Cartoon> cartoonList = IntStream.range(1, 31)
+        List<Cartoon> cartoonList = LongStream.range(1, 31)
                 .mapToObj(i -> Cartoon.builder()
                         .title("만화 제목 " + i)
                         .likes(i)
